backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
stripe
elevenlabs
pyttsx3
pydub
speechrecognition
httpx
jinja2
sqlalchemy
passlib[bcrypt]
email-validator
aiofiles
python-multipart
pytest
loguru
qrcode
pandas
openpyxl
requests
smtplib
